+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## v2.0.20.4
 - Avoid duplicate ad request when play the video
 - Pause video when preroll ad is shown and avoid showing video carousel
+- Fix duplicated ad request at first video play
 
 ## v2.0.20.3
 - Show player container before hiding poster